ldpd: move l2vpn API to ldp_l2vpn[_cli].[ch] files
The l2vpn structures are isolated from the rest of ldp code.
The main APIs used to handle pw, l2vpn if and l2vpn structure are moved
to ldp_l2vpn_cli.c file.
