+++ ldpd/ldpd.h
@@ -23,6 +23,7 @@
 #include "zclient.h"
 #include "nexthop.h"
 
+#include "ldp_l2vpn.h"
 #include "ldp.h"
 #include "lib/ldp_sync.h"
 
@@ -440,73 +441,6 @@ struct ldp_entity_stats {
 	uint32_t		 shutdown_rcv_notify;
 	uint32_t		 shutdown_send_notify;
 };
-
-struct l2vpn_if {
-	RB_ENTRY(l2vpn_if)	 entry;
-	struct l2vpn		*l2vpn;
-	char ifname[IFNAMSIZ];
-	ifindex_t		 ifindex;
-	int			 operative;
-	uint8_t			 mac[ETH_ALEN];
-	QOBJ_FIELDS;
-};
-RB_HEAD(l2vpn_if_head, l2vpn_if);
-RB_PROTOTYPE(l2vpn_if_head, l2vpn_if, entry, l2vpn_if_compare);
-DECLARE_QOBJ_TYPE(l2vpn_if);
-
-struct l2vpn_pw {
-	RB_ENTRY(l2vpn_pw)	 entry;
-	struct l2vpn		*l2vpn;
-	struct in_addr		 lsr_id;
-	int			 af;
-	union g_addr addr;
-	uint32_t		 pwid;
-	char ifname[IFNAMSIZ];
-	ifindex_t		 ifindex;
-	bool			 enabled;
-	uint32_t		 remote_group;
-	uint16_t		 remote_mtu;
-	uint32_t		 local_status;
-	uint32_t		 remote_status;
-	uint8_t			 flags;
-	uint8_t			 reason;
-	QOBJ_FIELDS;
-};
-RB_HEAD(l2vpn_pw_head, l2vpn_pw);
-RB_PROTOTYPE(l2vpn_pw_head, l2vpn_pw, entry, l2vpn_pw_compare);
-DECLARE_QOBJ_TYPE(l2vpn_pw);
-#define F_PW_STATUSTLV_CONF	0x01	/* status tlv configured */
-#define F_PW_STATUSTLV		0x02	/* status tlv negotiated */
-#define F_PW_CWORD_CONF		0x04	/* control word configured */
-#define F_PW_CWORD		0x08	/* control word negotiated */
-#define F_PW_STATIC_NBR_ADDR	0x10	/* static neighbor address configured */
-#define F_PW_SEND_REMOTE	0x20	/* send pw message to remote */
-
-
-#define F_PW_NO_ERR             0x00	/* no error reported */
-#define F_PW_LOCAL_NOT_FWD      0x01	/* locally can't forward over PW */
-#define F_PW_REMOTE_NOT_FWD     0x02	/* remote end of PW reported fwd error*/
-#define F_PW_NO_REMOTE_LABEL    0x03	/* have not recvd label from peer */
-#define F_PW_MTU_MISMATCH       0x04	/* mtu mismatch between peers */
-
-struct l2vpn {
-	RB_ENTRY(l2vpn)		 entry;
-	char			 name[L2VPN_NAME_LEN];
-	int			 type;
-	int			 pw_type;
-	int			 mtu;
-	char br_ifname[IFNAMSIZ];
-	ifindex_t		 br_ifindex;
-	struct l2vpn_if_head	 if_tree;
-	struct l2vpn_pw_head	 pw_tree;
-	struct l2vpn_pw_head	 pw_inactive_tree;
-	QOBJ_FIELDS;
-};
-RB_HEAD(l2vpn_head, l2vpn);
-RB_PROTOTYPE(l2vpn_head, l2vpn, entry, l2vpn_compare);
-DECLARE_QOBJ_TYPE(l2vpn);
-#define L2VPN_TYPE_VPWS		1
-#define L2VPN_TYPE_VPLS		2
 
 /* ldp_conf */
 extern enum ldpd_process {
